add smartstream ref
