+++ crates/fluvio-sc/Cargo.toml
@@ -48,7 +48,7 @@ fluvio-auth = { version = "0.0.0", path = "../fluvio-auth" }
 fluvio-future = { version = "0.3.0", features = ["subscriber", "openssl_tls", "zero_copy"] }
 fluvio-types = { version = "0.2.0", path = "../fluvio-types", features = ["events"] }
 fluvio-sc-schema = { version = "0.11.0", path = "../fluvio-sc-schema" }
-fluvio-stream-model = { version = "0.5.0", path = "../fluvio-stream-model" }
+fluvio-stream-model = { version = "0.5.4", path = "../fluvio-stream-model" }
 fluvio-controlplane = { version = "0.0.0", path = "../fluvio-controlplane" }
 fluvio-controlplane-metadata = { version = "0.12.0", features = ["k8", "serde"], path = "../fluvio-controlplane-metadata" }
 fluvio-stream-dispatcher = { version = "0.6.5", path = "../fluvio-stream-dispatcher" }
