+++ crates/fluvio-sc/src/controllers/smartstreams/controller.rs
@@ -103,51 +103,11 @@ where
 #[cfg(test)]
 mod test {
 
+    use fluvio_stream_model::store::memory::MemoryMeta;
+
     use crate::controllers::smartstreams;
 
     use super::*;
-
-    mod meta {
-        use fluvio_controlplane_metadata::core::{MetadataItem, MetadataRevExtension};
-
-        /// simple memory representation of meta
-        #[derive(Debug, Default, PartialEq, Clone)]
-        pub struct MemoryMeta {
-            pub rev: u32,
-        }
-
-        impl MetadataItem for MemoryMeta {
-            type UId = u32;
-
-            fn uid(&self) -> &Self::UId {
-                &self.rev
-            }
-
-            fn is_newer(&self, another: &Self) -> bool {
-                self.rev >= another.rev
-            }
-        }
-
-        impl MetadataRevExtension for MemoryMeta {
-            fn next_rev(&self) -> Self {
-                Self { rev: self.rev + 1 }
-            }
-        }
-
-        impl MemoryMeta {
-            pub fn new(rev: u32) -> Self {
-                Self { rev }
-            }
-        }
-
-        /*
-        impl From<u32> for MetadataContext<MemoryMeta> {
-            fn from(val: u32) -> MetadataContext<MemoryMeta> {
-                MemoryMeta::new(val).into()
-            }
-        }
-        */
-    }
 
     use fluvio_controlplane_metadata::{
         smartstream::{SmartStreamInput, SmartStreamInputs, SmartStreamRef},
@@ -156,7 +116,6 @@ mod test {
             actions::{LSChange, LSUpdate},
         },
     };
-    use meta::*;
 
     type MemSmartStreams = MetadataStoreObject<SmartStreamSpec, MemoryMeta>;
     type MemSmartStream = MetadataStoreObject<SmartStreamSpec, MemoryMeta>;
